requirements.txt
-i https://pypi.org/simple
aiofiles==24.1.0; python_version >= '3.8'
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.11.0; python_version >= '3.9'
async-timeout==5.0.1; python_full_version < '3.11.3'
certifi==2025.10.5; python_version >= '3.7'
cffi==2.0.0; python_full_version >= '3.9' and platform_python_implementation != 'PyPy'
charset-normalizer==3.4.3; python_version >= '3.7'
click==8.3.0; python_version >= '3.10'
colorama==0.4.6; platform_system == 'Windows'
cryptography==46.0.2; python_version >= '3.8' and python_full_version not in '3.9.0, 3.9.1'
distro==1.9.0; python_version >= '3.6'
faiss-cpu==1.12.0; python_version < '3.15' and python_version >= '3.9'
fastapi==0.118.0; python_version >= '3.8'
filelock==3.19.1; python_version >= '3.9'
fsspec==2025.9.0; python_version >= '3.9'
greenlet==3.2.4; python_version < '3.14' and platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
h11==0.16.0; python_version >= '3.8'
httpcore==1.0.9; python_version >= '3.8'
httpx==0.28.1; python_version >= '3.8'
huggingface-hub==0.35.3; python_full_version >= '3.8.0'
idna==3.10; python_version >= '3.6'
jinja2==3.1.6; python_version >= '3.7'
jiter==0.11.0; python_version >= '3.9'
joblib==1.5.2; python_version >= '3.9'
jsonpatch==1.33; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
jsonpointer==3.0.0; python_version >= '3.7'
langchain==0.3.27; python_version >= '3.9' and python_version < '4.0'
langchain-core==0.3.78; python_full_version >= '3.9.0' and python_full_version < '4.0.0'
langchain-text-splitters==0.3.11; python_version >= '3.9'
langsmith==0.4.32; python_version >= '3.9'
lxml==6.0.2; python_version >= '3.8'
markupsafe==3.0.3; python_version >= '3.9'
mpmath==1.3.0
networkx==3.5; python_version >= '3.11'
numpy==2.3.3; python_version >= '3.11'
openai==2.2.0; python_version >= '3.8'
orjson==3.11.3; platform_python_implementation != 'PyPy'
packaging==25.0; python_version >= '3.8'
pdfminer.six==20250506; python_version >= '3.9'
pdfplumber==0.11.7; python_version >= '3.8'
pillow==11.3.0; python_version >= '3.9'
pycparser==2.23; implementation_name != 'PyPy'
pydantic==2.12.0; python_version >= '3.9'
pydantic-core==2.41.1; python_version >= '3.9'
pymupdf==1.26.4; python_version >= '3.9'
pypdf==6.1.1; python_version >= '3.9'
pypdfium2==4.30.0; python_version >= '3.6'
python-docx==1.2.0; python_version >= '3.9'
python-multipart==0.0.20; python_version >= '3.8'
pyyaml==6.0.3; python_version >= '3.8'
rank-bm25==0.2.2
redis==6.4.0; python_version >= '3.9'
regex==2025.9.18; python_version >= '3.9'
requests==2.32.5; python_version >= '3.9'
requests-toolbelt==1.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
safetensors==0.6.2; python_version >= '3.9'
scikit-learn==1.7.2; python_version >= '3.10'
scipy==1.16.2; python_version >= '3.11'
sentence-transformers==5.1.1; python_version >= '3.9'
sniffio==1.3.1; python_version >= '3.7'
sqlalchemy==2.0.43; python_version >= '3.7'
starlette==0.48.0; python_version >= '3.9'
sympy==1.14.0; python_version >= '3.9'
tenacity==9.1.2; python_version >= '3.9'
threadpoolctl==3.6.0; python_version >= '3.9'
tokenizers==0.22.1; python_version >= '3.9'
torch==2.8.0; python_full_version >= '3.9.0'
tqdm==4.67.1; python_version >= '3.7'
transformers==4.57.0; python_full_version >= '3.9.0'
typing-extensions==4.15.0; python_version >= '3.9'
typing-inspection==0.4.2; python_version >= '3.9'
urllib3==2.5.0; python_version >= '3.9'
uvicorn==0.37.0; python_version >= '3.9'
zstandard==0.25.0; python_version >= '3.9'
